+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.6.dev0
 
-- n/a
+- added `upload_fileobj()` to stream a file-like object to the bucket
+- added `download_fileobj()` to download to a file-like object as well
 
 # 0.5